+++ base-coroutine/src/lib.rs
@@ -18,6 +18,8 @@ pub use coroutine::*;
 #[allow(dead_code)]
 mod work_steal;
 
+pub use work_steal::*;
+
 #[allow(dead_code)]
 pub mod scheduler;
